Try to fix introduces to cycle in the graph exception Move some classes Improve some code
