+++ cloudnet-cord/cloudnet-setup/pom.xml
@@ -21,9 +21,10 @@
             <scope>compile</scope>
         </dependency>
         <dependency>
-            <groupId>de.dytanic.cloudnet</groupId>
-            <artifactId>cloudnet-wrapper</artifactId>
-            <version>${project.version}</version>
+            <!-- jsoup HTML parser library @ https://jsoup.org/ -->
+            <groupId>org.jsoup</groupId>
+            <artifactId>jsoup</artifactId>
+            <version>1.12.1</version>
             <scope>compile</scope>
         </dependency>
     </dependencies>
