+++ cloudnet-cord/cloudnet-setup/src/main/java/de/dytanic/cloudnet/setup/models/PaperMCProjectVersion.java
@@ -1,4 +1,4 @@
-package de.dytanic.cloudnetwrapper.models;
+package de.dytanic.cloudnet.setup.models;
 
 public class PaperMCProjectVersion {
