+++ cloudnet-wrapper/src/main/java/de/dytanic/cloudnetwrapper/CloudNetWrapper.java
@@ -35,10 +35,10 @@ import de.dytanic.cloudnetwrapper.server.GameServer;
 import de.dytanic.cloudnetwrapper.server.process.ServerProcessQueue;
 import de.dytanic.cloudnet.setup.spigot.SetupSpigotVersion;
 import de.dytanic.cloudnetwrapper.util.FileUtility;
-import de.dytanic.cloudnetwrapper.util.PaperBuilder;
+import de.dytanic.cloudnet.setup.spigot.PaperBuilder;
 import de.dytanic.cloudnetwrapper.util.ShutdownHook;
 import de.dytanic.cloudnetwrapper.util.ShutdownOnCentral;
-import de.dytanic.cloudnetwrapper.util.SpigotBuilder;
+import de.dytanic.cloudnet.setup.spigot.SpigotBuilder;
 import joptsimple.OptionSet;
 import lombok.Getter;
 import lombok.Setter;
