+++ cloudnet-wrapper/src/main/java/de/dytanic/cloudnetwrapper/util/ShutdownHook.java
@@ -1,5 +1,7 @@
 package de.dytanic.cloudnetwrapper.util;
 
+import de.dytanic.cloudnet.setup.spigot.PaperBuilder;
+import de.dytanic.cloudnet.setup.spigot.SpigotBuilder;
 import de.dytanic.cloudnetwrapper.CloudNetWrapper;
 
 public class ShutdownHook implements Runnable {
